Fix: side of heat exchanger is lost.
- Extract `Side` class from `storageTank/model.py`.
- Remove top-level `side.py`: very easy to confuse these `Side` classes.
