--- trnsysGUI/side.py
@@ -1,6 +1,0 @@
-import enum as _enum
-
-
-class Side(_enum.Enum):
-    LEFT = "Left"
-    RIGHT = "Right"
